issues #64 technical debt

+++ api/controllers/common.go
@@ -134,7 +134,7 @@ func (c ControllerCommon) error(ctx context.Context, errs validator.ValidationEr
 	}
 }
 
-func (c ControllerCommon) Pagination(limit, offset uint32, order, sortBy string) (pagination common.PageParams) {
+func (c ControllerCommon) Pagination(limit, offset uint64, order, sortBy string) (pagination common.PageParams) {
 
 	pagination = common.PageParams{
 		Limit:  200,
