+++ api/controllers/image.go
@@ -79,7 +79,7 @@ func (c ControllerImage) GetImageList(ctx context.Context, req *api.GetImageList
 		return nil, c.error(ctx, nil, err)
 	}
 
-	return c.dto.Image.ToImageListResult(items, uint32(total), req.Limit, req.Offset), nil
+	return c.dto.Image.ToImageListResult(items, uint64(total), req.Limit, req.Offset), nil
 }
 
 // DeleteImageById ...
