+++ api/controllers/script.go
@@ -78,7 +78,7 @@ func (c ControllerScript) GetScriptList(ctx context.Context, req *api.GetScriptL
 		return nil, c.error(ctx, nil, err)
 	}
 
-	return c.dto.Script.ToListResult(items, uint32(total), req.Limit, req.Offset), nil
+	return c.dto.Script.ToListResult(items, uint64(total), req.Limit, req.Offset), nil
 }
 
 // SearchScriptById ...
