+++ api/dto/plugin.go
@@ -33,7 +33,7 @@ func NewPluginDto() Plugin {
 }
 
 // ToPluginListResult ...
-func (p Plugin) ToPluginListResult(items []*m.Plugin, total, limit, offset uint32) (result *api.GetPluginListResult) {
+func (p Plugin) ToPluginListResult(items []*m.Plugin, total, limit, offset uint64) (result *api.GetPluginListResult) {
 
 	result = &api.GetPluginListResult{
 		Items: make([]*api.Plugin, 0, len(items)),
